Add `debounce()` & `throttle()` functions
Deprecates `debounce()` from `events.js` and just passes onto new functions from `utility.js`.
